+++ client/deps.go
@@ -30,6 +30,12 @@ type connectionChecker interface {
 	Check(ctx context.Context, req *agentpb.CheckConnectionRequest, id uint32) *agentpb.CheckConnectionResponse
 }
 
+// connectionChecker is a subset of methods of connectionchecker.ConnectionChecker used by this package.
+// We use it instead of real type for testing and to avoid dependency cycle.
+type connectionChecker interface {
+	Check(ctx context.Context, req *agentpb.CheckConnectionRequest, id uint32) *agentpb.CheckConnectionResponse
+}
+
 // supervisor is a subset of methods of supervisor.Supervisor used by this package.
 // We use it instead of real type for testing and to avoid dependency cycle.
 type supervisor interface {
